split up swarm.cpp and shrink the number of views that are pre-instantiated.
